backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
boto3==1.34.0
httpx==0.25.2
python-dotenv==1.0.0
google-generativeai==0.3.2
